split propagator from md
